Oc 1073 smt remove gson

+++ services/core/cards-publication/src/main/java/org/lfenergy/operatorfabric/cards/publication/kafka/command/DeleteCardCommandHandler.java
@@ -1,14 +1,24 @@
 package org.lfenergy.operatorfabric.cards.publication.kafka.command;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.lfenergy.operatorfabric.avro.CardCommand;
 import org.lfenergy.operatorfabric.avro.CommandType;
 import org.lfenergy.operatorfabric.cards.publication.model.CardPublicationData;
+import org.lfenergy.operatorfabric.cards.publication.services.CardProcessingService;
 import org.springframework.stereotype.Component;
 
 @Component
 @Slf4j
 public class DeleteCardCommandHandler extends BaseCommandHandler implements CommandHandler {
+
+    private final CardProcessingService cardProcessingService;
+
+    public DeleteCardCommandHandler(CardProcessingService cardProcessingService, ObjectMapper objectMapper) {
+        super(objectMapper);
+        this.cardProcessingService = cardProcessingService;
+    }
 
     @Override
     public CommandType getCommandType() {
@@ -21,6 +31,8 @@ public class DeleteCardCommandHandler extends BaseCommandHandler implements Comm
                 cardCommand.getProcessInstanceId(), cardCommand.getProcess(), cardCommand.getCard().getData());
 
         CardPublicationData card = buildCardPublicationData(cardCommand);
-        cardProcessingService.deleteCard(card.getProcessInstanceId());
+        if (card != null) {
+            cardProcessingService.deleteCard(card.getProcessInstanceId());
+        }
     }
 }
